+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Auto compression option is part of drop down (no extra checkbox) ([#160](https://github.com/getsentry/sentry-unity/pull/160))
 - Rename DiagnosticsLogger to DiagnosticLogger ([#168](https://github.com/getsentry/sentry-unity/pull/168))
 - SentryOptions config proper check ([#176](https://github.com/getsentry/sentry-unity/pull/176))
+- Diagnostic logger writes to console that it was disabled ([#183](https://github.com/getsentry/sentry-unity/pull/183))
 
 ## 0.0.13