+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.0.14
 
 - il2cpp remove zeroes from path ([#179](https://github.com/getsentry/sentry-unity/pull/179))
 - Flag simulator based on Application.isEditor ([#184](https://github.com/getsentry/sentry-unity/pull/184))